feat: add benchmarks and optimization code
