+++ src/OtherMonad.Maybe/Maybe.Bind.Defer.cs
@@ -1,5 +1,7 @@
 ﻿namespace OtherMonad;
 
+using System.Runtime.CompilerServices;
+
 /// <summary>
 /// Extension methods to Maybe Monad
 /// </summary>
@@ -14,6 +16,7 @@ public static partial class Maybe
     /// <param name="selector">A transform function to apply to source element</param>
     /// <returns><see cref="Deferred{Maybe}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>>]]></see></returns>
     /// <exception cref="ArgumentNullException">selector is null</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TResult>> BindDefer<TSource, TResult>(this Maybe<TSource> source, Func<TSource, Maybe<TResult>> selector)
     {
         ArgumentNullException.ThrowIfNull(selector);
@@ -30,6 +33,7 @@ public static partial class Maybe
     /// <param name="selector">A transform function to apply to source element</param>
     /// <returns><see cref="Deferred{Maybe}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>>]]></see></returns>
     /// <exception cref="ArgumentNullException">selector is null</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TResult>> BindDefer<TSource, TResult>(this Deferred<Maybe<TSource>> source, Func<TSource, Maybe<TResult>> selector)
     {
         ArgumentNullException.ThrowIfNull(selector);
@@ -51,6 +55,7 @@ public static partial class Maybe
     /// <param name="cancellation">A CancellationToken enables cooperative cancellation between threads, thread pool work items, or Task objects</param>
     /// <returns><see cref="DeferredTask{Maybe}"><![CDATA[ DeferredTask<Maybe<]]><typeparamref name="TResult"/><![CDATA[>>]]></see></returns>
     /// <exception cref="ArgumentNullException">selector is null</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static DeferredTask<Maybe<TResult>> BindDefer<TSource, TResult>(this Maybe<TSource> source, Func<TSource, CancellationToken, Task<Maybe<TResult>>> selector, CancellationToken cancellation = default)
     {
         ArgumentNullException.ThrowIfNull(selector);
@@ -68,6 +73,7 @@ public static partial class Maybe
     /// <param name="cancellation">A CancellationToken enables cooperative cancellation between threads, thread pool work items, or Task objects</param>
     /// <returns><see cref="DeferredTask{Maybe}"><![CDATA[ DeferredTask<Maybe<]]><typeparamref name="TResult"/><![CDATA[>>]]></see></returns>
     /// <exception cref="ArgumentNullException">selector is null</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static DeferredTask<Maybe<TResult>> BindDefer<TSource, TResult>(this DeferredTask<Maybe<TSource>> source, Func<TSource, CancellationToken, Task<Maybe<TResult>>> selector, CancellationToken cancellation = default)
     {
         ArgumentNullException.ThrowIfNull(selector);
