+++ src/OtherMonad.Maybe/Maybe.Combine.Defer.cs
@@ -1,5 +1,7 @@
 ﻿namespace OtherMonad;
 
+using System.Runtime.CompilerServices;
+
 /// <summary>
 /// Extension methods to Maybe Monad
 /// </summary>
@@ -16,6 +18,7 @@ public static partial class Maybe
     /// <param name="other">A right value to invoke a combine</param>
     /// <param name="select">A combine function to apply to source element with other</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TResult>> CombineDefer<TSource, TCombine, TResult>(this Deferred<Maybe<TSource>> source, Maybe<TCombine> other, Func<TSource, TCombine, TResult> select)
     {
         return () =>
@@ -39,6 +42,7 @@ public static partial class Maybe
     /// <param name="other">A right value to invoke a combine</param>
     /// <param name="select">A combine function to apply to source element with other</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TResult>> CombineDefer<TSource, TCombine, TResult>(this Deferred<Maybe<TSource>> source, Deferred<Maybe<TCombine>> other, Func<TSource, TCombine, TResult> select)
     {
         return () =>
@@ -65,6 +69,7 @@ public static partial class Maybe
     /// <param name="select">A combine function to apply to source element with other</param>
     /// <param name="cancellation">A CancellationToken enables cooperative cancellation between threads, thread pool work items, or Task objects</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static DeferredTask<Maybe<TResult>> CombineDefer<TSource, TCombine, TResult>(this DeferredTask<Maybe<TSource>> source, Maybe<TCombine> other, Func<TSource, TCombine, CancellationToken, Task<TResult>> select, CancellationToken cancellation = default)
     {
         return async () =>
@@ -89,6 +94,7 @@ public static partial class Maybe
     /// <param name="select">A combine function to apply to source element with other</param>
     /// <param name="cancellation">A CancellationToken enables cooperative cancellation between threads, thread pool work items, or Task objects</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static DeferredTask<Maybe<TResult>> CombineDefer<TSource, TCombine, TResult>(this DeferredTask<Maybe<TSource>> source, DeferredTask<Maybe<TCombine>> other, Func<TSource, TCombine, CancellationToken, Task<TResult>> select, CancellationToken cancellation = default)
     {
         return async () =>
@@ -115,6 +121,7 @@ public static partial class Maybe
     /// <param name="select">A combine function to apply to source element with other</param>
     /// <param name="defaultValueFactory">Default value in case no match</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TResult>> TryCombineDefer<TSource, TCombine, TResult>(this Deferred<Maybe<TSource>> source, Maybe<TCombine> other, Func<TSource, TCombine, TResult> select, Func<TResult> defaultValueFactory)
     {
         return () =>
@@ -144,6 +151,7 @@ public static partial class Maybe
     /// <param name="select">A combine function to apply to source element with other</param>
     /// <param name="defaultValueFactory">Default value in case no match</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ Deferred<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TResult>> TryCombineDefer<TSource, TCombine, TResult>(this Deferred<Maybe<TSource>> source, Deferred<Maybe<TCombine>> other, Func<TSource, TCombine, TResult> select, Func<TResult> defaultValueFactory)
     {
         return () =>
@@ -174,6 +182,7 @@ public static partial class Maybe
     /// <param name="defaultValueFactory">Default value in case no match</param>
     /// <param name="cancellation">A CancellationToken enables cooperative cancellation between threads, thread pool work items, or Task objects</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ DeferredTask<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static DeferredTask<Maybe<TResult>> TryCombineDefer<TSource, TCombine, TResult>(this DeferredTask<Maybe<TSource>> source, Maybe<TCombine> other, Func<TSource, TCombine, CancellationToken, Task<TResult>> select, Func<Task<TResult>> defaultValueFactory, CancellationToken cancellation = default)
     {
         return async () =>
@@ -204,6 +213,7 @@ public static partial class Maybe
     /// <param name="defaultValueFactory">Default value in case no match</param>
     /// <param name="cancellation">A CancellationToken enables cooperative cancellation between threads, thread pool work items, or Task objects</param>
     /// <returns><see cref="Maybe{TResult}"><![CDATA[ DeferredTask<Maybe<]]><typeparamref name="TResult"/><![CDATA[>> ]]></see> or default value</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static DeferredTask<Maybe<TResult>> TryCombineDefer<TSource, TCombine, TResult>(this DeferredTask<Maybe<TSource>> source, DeferredTask<Maybe<TCombine>> other, Func<TSource, TCombine, CancellationToken, Task<TResult>> select, Func<Task<TResult>> defaultValueFactory, CancellationToken cancellation = default)
     {
         return async () =>
