+++ src/OtherMonad.Maybe/Maybe.OrElse.Defer.cs
@@ -1,5 +1,7 @@
 ﻿namespace OtherMonad;
 
+using System.Runtime.CompilerServices;
+
 /// <summary>
 /// Extension methods to Maybe Monad
 /// </summary>
@@ -12,6 +14,7 @@ public static partial class Maybe
     /// <param name="source">A value to invoke to check if has value</param>
     /// <param name="default">Default value to return if dont has value</param>
     /// <returns><see cref="Deferred{Maybe}"><![CDATA[Deferred<Maybe<]]><typeparamref name="TSource"/><![CDATA[>>]]></see></returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TSource>> OrElseDefer<TSource>(this Maybe<TSource> source, TSource @default)
     {
         return () => source.OrElse(@default);
@@ -24,6 +27,7 @@ public static partial class Maybe
     /// <param name="source">A value to invoke to check if has value</param>
     /// <param name="default">Default value to return if dont has value</param>
     /// <returns><see cref="Deferred{Maybe}"><![CDATA[Deferred<Maybe<]]><typeparamref name="TSource"/><![CDATA[>>]]></see></returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Deferred<Maybe<TSource>> OrElseDefer<TSource>(this Deferred<Maybe<TSource>> source, TSource @default)
     {
         return () =>
@@ -40,6 +44,7 @@ public static partial class Maybe
     /// <param name="source">A value to invoke to check if has value</param>
     /// <param name="default">Default value to return if dont has value</param>
     /// <returns><see cref="DeferredTask{Maybe}"><![CDATA[DeferredTask<Maybe<]]><typeparamref name="TSource"/><![CDATA[>>]]></see></returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static DeferredTask<Maybe<TSource>> OrElseDefer<TSource>(this DeferredTask<Maybe<TSource>> source, TSource @default)
     {
         return async () =>
