+++ src/OtherMonad.Maybe/Maybe.Wrap.cs
@@ -1,5 +1,7 @@
 ﻿namespace OtherMonad;
 
+using System.Runtime.CompilerServices;
+
 /// <summary>
 /// Extension methods to Maybe Monad
 /// </summary>
@@ -11,6 +13,7 @@ public static partial class Maybe
     /// <typeparam name="TSource">The type of the value to wrap.</typeparam>
     /// <param name="source">The value to wrap. If <see langword="null"/>, returns <see cref="Maybe{TSource}.None"/>.</param>
     /// <returns>A <see cref="Maybe{TSource}"/> containing the value.</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Maybe<TSource> Wrap<TSource>(this TSource source)
     {
         return source;
@@ -23,6 +26,7 @@ public static partial class Maybe
     /// <param name="source">The <see cref="Maybe{TSource}"/> to unwrap.</param>
     /// <returns>The contained value.</returns>
     /// <exception cref="InvalidOperationException">Thrown when <paramref name="source"/> has no value (<see cref="Maybe{TSource}.HasValue"/> is <see langword="false"/>). Use <see cref="Unwrap{TSource}(Maybe{TSource}, TSource)"/> to provide a fallback value.</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static TSource Unwrap<TSource>(this Maybe<TSource> source)
     {
         if (!source.HasValue)
@@ -38,6 +42,7 @@ public static partial class Maybe
     /// <param name="source">The <see cref="Maybe{TSource}"/> to unwrap.</param>
     /// <param name="default">The default value to return if the <see cref="Maybe{TSource}"/> has no value.</param>
     /// <returns>The contained value if present; otherwise, <paramref name="default"/>.</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static TSource Unwrap<TSource>(this Maybe<TSource> source, TSource @default)
     {
         if (source.HasValue)
